fix: routes names, login & logout through button in appbar, new main page
