+++ src/container/MainPage.tsx
@@ -1,0 +1,12 @@
+function MainPage() {
+    return (
+         <div className='flex flex-col justify-center text-center items-center gap-4'>
+            <div className='font-bold'>
+                <h1>Welcome to Tomeiro Proyect</h1>
+            </div>
+            <img src="public/tomeiro_favicon.png" alt="tomeiro logo" width={200} height={200} />
+        </div>
+    )
+}
+
+export default MainPage
