+++ src/routeTree.gen.ts
@@ -9,68 +9,68 @@
 // Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
 
 import { Route as rootRouteImport } from './routes/__root'
-import { Route as WelcomePageRouteImport } from './routes/WelcomePage'
-import { Route as LoginPageRouteImport } from './routes/LoginPage'
+import { Route as LoginRouteImport } from './routes/login'
+import { Route as IndexRouteImport } from './routes/index'
 
-const WelcomePageRoute = WelcomePageRouteImport.update({
-  id: '/WelcomePage',
-  path: '/WelcomePage',
+const LoginRoute = LoginRouteImport.update({
+  id: '/login',
+  path: '/login',
   getParentRoute: () => rootRouteImport,
 } as any)
-const LoginPageRoute = LoginPageRouteImport.update({
-  id: '/LoginPage',
-  path: '/LoginPage',
+const IndexRoute = IndexRouteImport.update({
+  id: '/',
+  path: '/',
   getParentRoute: () => rootRouteImport,
 } as any)
 
 export interface FileRoutesByFullPath {
-  '/LoginPage': typeof LoginPageRoute
-  '/WelcomePage': typeof WelcomePageRoute
+  '/': typeof IndexRoute
+  '/login': typeof LoginRoute
 }
 export interface FileRoutesByTo {
-  '/LoginPage': typeof LoginPageRoute
-  '/WelcomePage': typeof WelcomePageRoute
+  '/': typeof IndexRoute
+  '/login': typeof LoginRoute
 }
 export interface FileRoutesById {
   __root__: typeof rootRouteImport
-  '/LoginPage': typeof LoginPageRoute
-  '/WelcomePage': typeof WelcomePageRoute
+  '/': typeof IndexRoute
+  '/login': typeof LoginRoute
 }
 export interface FileRouteTypes {
   fileRoutesByFullPath: FileRoutesByFullPath
-  fullPaths: '/LoginPage' | '/WelcomePage'
+  fullPaths: '/' | '/login'
   fileRoutesByTo: FileRoutesByTo
-  to: '/LoginPage' | '/WelcomePage'
-  id: '__root__' | '/LoginPage' | '/WelcomePage'
+  to: '/' | '/login'
+  id: '__root__' | '/' | '/login'
   fileRoutesById: FileRoutesById
 }
 export interface RootRouteChildren {
-  LoginPageRoute: typeof LoginPageRoute
-  WelcomePageRoute: typeof WelcomePageRoute
+  IndexRoute: typeof IndexRoute
+  LoginRoute: typeof LoginRoute
 }
 
 declare module '@tanstack/react-router' {
   interface FileRoutesByPath {
-    '/WelcomePage': {
-      id: '/WelcomePage'
-      path: '/WelcomePage'
-      fullPath: '/WelcomePage'
-      preLoaderRoute: typeof WelcomePageRouteImport
+    '/login': {
+      id: '/login'
+      path: '/login'
+      fullPath: '/login'
+      preLoaderRoute: typeof LoginRouteImport
       parentRoute: typeof rootRouteImport
     }
-    '/LoginPage': {
-      id: '/LoginPage'
-      path: '/LoginPage'
-      fullPath: '/LoginPage'
-      preLoaderRoute: typeof LoginPageRouteImport
+    '/': {
+      id: '/'
+      path: '/'
+      fullPath: '/'
+      preLoaderRoute: typeof IndexRouteImport
       parentRoute: typeof rootRouteImport
     }
   }
 }
 
 const rootRouteChildren: RootRouteChildren = {
-  LoginPageRoute: LoginPageRoute,
-  WelcomePageRoute: WelcomePageRoute,
+  IndexRoute: IndexRoute,
+  LoginRoute: LoginRoute,
 }
 export const routeTree = rootRouteImport
   ._addFileChildren(rootRouteChildren)
