+++ src/routes/index.tsx
@@ -1,0 +1,6 @@
+import MainPage from '@/container/MainPage'
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute('/')({
+    component: MainPage,
+})
